requirements.txt
# Biology RAG Chatbot - Requirements

# Core dependencies
streamlit>=1.30.0
google-generativeai>=0.3.0

# RAG components
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
langchain-text-splitters>=0.0.1

# PDF processing
pypdf>=3.17.0

# Scientific computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
